fix(colorschemes): not load colorschemes if passed in.
Well, this is a breaking change and I'm not sure if you will like it or
not.

What I want to do is to create a custom command to show a selected list
of colorschemes. It is impossible to achieve without introducing this
breaking change, or adding a new parameter.

I think the filtering behavior is easier to use. If people want to
extend the list, they can always use `vim.fn.getcompletion` themselves.