feat(builtin previewer): toggle behavior "extend" option (#721)
by default the main window dimensions account for the previewer
size, when hiding the preview the main window dimensions will
be modified to occupy the maximum size as if both the main window
and the previewer are drawn, similar to how fzf works in the shell.

By setting `previewers.builtin.toggle_behavior = "extend"` the size
of the main window will always stay the same, i.e. the previewer
"extends" the main window.