feat(input): Use asynchronous vim.ui.input with coroutine
This is an attempt continued from 6ee73fd, trying to use `vim.ui.input`
for asking user inputs via `fzf-lua.utils.input()`.

Previously in commit 6ee73fd (later reverted in 5d8dd50), the use of
`vim.ui.input` was incorrect, because it's an asynchronous function
but `utils.input()` is meant to be a synchronous function.
When dressing.nvim is used for `vim.ui.input`, `utils.input()` will
immediately return with `nil` before user enters text in the prompt;
it would result in the grep command being cancelled.

A possible fix to this would be to use a coroutine, which can "wrap"
an asynchronous function so that it can be used like a synchronous
function. To make it possible, command functions (e.g., FzfLua grep)
that uses `utils.input()` inside should be wrapped with coroutines.
Note that `utils.input()` will fall back to the default behavior
with the synchronous `vim.fn.input()` when coroutine is unavailable.

We enable the use of `vim.ui.input()` for prompting an user input
in neovim 0.9.0+ or higher, because the behavior of `vim.ui.input()`
on aborting the input by 'Ctrl-C' has been changed since neovim 0.9.0.