fix(core)!: always shellescape fzf_opts["--header"]
It was previously not clearly documented or required whether
`fzf_opts["--header"]` must be shell-escaped or may take a form of raw
strings.

BREAKING CHANGE: Like `fzf_opts["--prompt"]`, `fzf_opts["--header"]`
will be automatically shell-escaped (for the sake of consistency)
so normal string can be directly fed to `opts.fzf_opts["--header"]`.

Although this may not break any built-in commands, any user-custom
opts that have a *shell-escaped* header fzf_opts may result in a wrong
double escaping.

This commit also adds an unit test for this behavior on `--header`.