fix(git_status): untracked files should be displayed correctly
Problem: The git status of untracked files in an (untracked)
directory were not correctly recognized as '?'.

Solution: Search upward in path to see if any of the parent or ancestor
directory is marked as untracked by `git status`.

Alternatively, we considered using `-u` (`--untracked-files`) option
for git status, but this may enumerate too many (untracked) files in the
directory which may have a performance problem.

Note:

- Use git status --porcelain command, this is always a preferred
way to generate machine-readable git status.

- Also, the logic for git status parsing has been improved.

Refs: https://git-scm.com/docs/git-status