refactor(utils)!: use vim.system for running process if available
Use the new `vim.system()` API on neovim 0.10 or higher, in place of
`vim.fn.system()`. It's more performant by using libuv (`vim.loop`),
though only slightly, and is known to have less glitches.

The experimental feature `lua_io` is removed, as this was actually not
being used; normal users are not expected to have been using this
(undocumented) feature.