fix(manpages): preview and open the correct section for the man page
The man pages are grouped into different sections and it is common for
identically named pages to be found in multiple sections. For instance
`printf(1)` documenting the shell command and `printf(3)` documenting
the C-function.

Prior to these changes the section name was stripped away from the
selected entries, so it only previews and opens that page from the
default section even though another section was selected. Now the
apropos entry is rewritten to a `:Man` argument with the section name.