fix(preview): set flex threshold from window columns in split mode
The split mode opens the fzf selection in a split from the current
window, so the number of available columns can be less than the width of
the full terminal, which should be taken into the account when
calculating whether the width exceeds the flex threshold or not.

These changes also deduplicate some of the flex logic shared between
native mode (in core.lua) and builtin mode (in win.lua).