fix(quickfix): handle multi-line entries more like the quicklist window
That means having them displayed without new lines so that they are
concatenated into on one long line. Prior to these changes if a quickfix
entry included newlines those would become separate entries in the fzf
selection list. Quickfix entries are formatted so they start with the
file name and location, but only the first line is like that, so if one
were to select one of the other lines from an entry the start of this
line would be interpreted as a file name and would open a buffer "there"