fix(ctag): better handle line number for tag entries
Tag entries with no regex did not properly extract the line number, thus
were unusable.
Tag entries with both a regex and a line number did not take advantage
of the line number, thus were susceptible to collisions on the regex.
The default previewer for tags could not use the line number of an entry