fix(icons): restore auto lazy-load (closes #1670)
This commits restores lazy loading the icon package albeit with
a slight difference, the defaults now set `file_icons = 1` which
is similar to `file_icons = true` (= trigger lazy loading) without
displaying a warning message if we are unable to load the package.

This change does not affect #1663, fzf-lua will not trigger lazy
loading of the icon package when `file_icons = false`.