feat(profiles): added `hide` (improved resume, #1686)
The profile changes the default binds to hide the fzf-lua window
instead of aborting it, therefore keeping cursor position, selection,
etc.

Note that this also continues long running operations in the background
so be mindful of sending fzf to background on a large mono repo, can
still be aborted with `ctrl-c` or `alt-esc`.

Enable with:
```lua
-- Set as base profile
:lua require("fzf-lua").setup({"hide"})
-- More than one profile
:lua require("fzf-lua").setup({"border-fused","hide"})
-- Or with `profiles`:
:FzfLua profiles load=hide
-- More than one profile
:FzfLua profiles load={"border-fused","hide"}
```