fix(code_actions)!: do not query LSP directly (closes #2044)
FzfLua will no longer pre-query the LSP to see if there are code
actions available when we aren't registered as vim.ui.select backend.
Instead, a warning will be displayed to the user, fzf-lua will still
attempt to register and de-register in the exit/accpept callback but
with `hide` profile or when no code actions are available fzf-lua will
remain the registered ui.select interface until `deregister_ui_select`
is called.

This is much less error prone than having to keep up with upstream
changes of the `vim.lsp.buf.code_action()` works.