fix(reflow): resume causing lines to shift
## Details

This change disables `FileType` events from triggering on the backdrop
buffer. Why exactly this causes issues in the terminal buffer, I'm not
sure. My best guess is that having a variety of plugins running
autocommands for the backdrop causes some timing issue, maybe in the
redraw cycle. I believe this solves my issue and potentially others:

- https://github.com/ibhagwan/fzf-lua/issues/2025
- https://github.com/ibhagwan/fzf-lua/discussions/2457