feat(git_diff): add compare_against option, diff against parent by default
An additional field is added to `git_diff()`: `compare_against` (string).
If it is not provided, behaviour is also changed to diff against the
provided `ref` and its direct parent commit, showing what changes the `ref`
commit introduced. When `compare_against` is given, the difference between
commits `compare_against` and `ref` is shown. The `git_preview()` function
is modified to operate from the nearest Git repository up the `cwd` path,
so that the preview for `git_diff()` works.